fix(*): fix rpc auth

+++ ui/helpers/angular.helpers.ts
@@ -2,9 +2,12 @@ import { OnDestroy, OnInit } from '@angular/core'
 import { Subject } from 'rxjs'
 import { takeUntil } from 'rxjs/operators'
 
-export function MortalityAware(): ClassDecorator {
-  return (target: any) => {
-    if (!target.prototype.qb__onDestroy$) {
+export function MortalityAware(): ClassDecorator
+{
+  return (target: any) =>
+  {
+    if (!target.prototype.qb__onDestroy$)
+    {
       const originalOnDestroy = typeof target.prototype.ngOnDestroy === 'function'
         ? target.prototype.ngOnDestroy.bind(target)
         : null
@@ -21,7 +24,8 @@ export function MortalityAware(): ClassDecorator {
   }
 }
 
-export const forLifeOf = <ValueType>(instance: any) => {
+export const forLifeOf = <ValueType>(instance: any) =>
+{
   // TODO: Check to make sure the class is a component class.
   // (http://prideparrot.com/blog/archive/2018/7/extending_component_decorator_angular_6)
   // Check to make sure the class has the @MortalityAwareComponent decorator.
@@ -33,22 +37,31 @@ export const forLifeOf = <ValueType>(instance: any) => {
   return takeUntil<ValueType>(instance.qb__onDestroy$)
 }
 
-export interface ModuleApi {
+export interface ModuleApi
+{
   init: () => any
   destroy?: () => any
 }
 
-export abstract class ModuleContainer implements OnInit, OnDestroy {
+export abstract class ModuleContainer implements OnInit, OnDestroy
+{
   constructor(
     public api: ModuleApi
-  ) { }
+  )
+  { }
 
-  public ngOnInit(): void {
-    this.api.init()
+  public ngOnInit(): void
+  {
+    if (this.api)
+    {
+      this.api.init()
+    }
   }
 
-  public ngOnDestroy(): void {
-    if (this.api.destroy) {
+  public ngOnDestroy(): void
+  {
+    if (this.api && this.api.destroy)
+    {
       this.api.destroy()
     }
   }
